+++ dxsp/utils/__init__.py
@@ -1,7 +1,9 @@
 from dxsp.utils.account_utils import AccountUtils
 from dxsp.utils.contract_utils import ContractUtils
+from dxsp.utils.token_utils import Token
 
 __all__ = [
     "AccountUtils",
     "ContractUtils",
+    "Token",
 ]
